+++ dakara_server/playlist/views.py
@@ -76,7 +76,8 @@ class PlaylistEntryListView(drf_generics.ListCreateAPIView):
 
     def get(self, request, *args, **kwargs):
         queryset = self.queryset.all()
-        player = models.Player.get_or_create()
+        karaoke = models.Karaoke.objects.get_object()
+        player, _ = models.Player.cache.get_or_create(karaoke=karaoke)
         date = datetime.now(tz)
 
         # add player remaining time
@@ -130,7 +131,7 @@ class PlaylistEntryListView(drf_generics.ListCreateAPIView):
         ):
             # compute playlist end date
             playlist = self.filter_queryset(self.get_queryset())
-            player = models.Player.get_or_create()
+            player, _ = models.Player.cache.get_or_create(karaoke=karaoke)
             date = datetime.now(tz)
 
             # add player remaining time
@@ -161,7 +162,7 @@ class PlaylistEntryListView(drf_generics.ListCreateAPIView):
         #   - player is set to play next song
         #   - the player is idle.
         next_playlist_entry = models.PlaylistEntry.objects.get_next()
-        player = models.Player.get_or_create()
+        player, _ = models.Player.cache.get_or_create(karaoke=karaoke)
         if all(
             (
                 next_playlist_entry is not None,
@@ -205,7 +206,7 @@ class PlayerCommandView(drf_generics.UpdateAPIView):
             )
 
         # check the player is not idle
-        player = models.Player.get_or_create()
+        player, _ = models.Player.cache.get_or_create(karaoke=karaoke)
         if player.playlist_entry is None:
             raise PermissionDenied("The player cannot receive commands when " "idle")
 
@@ -229,20 +230,21 @@ class DigestView(APIView):
 
     def get(self, request, *args, **kwargs):
         """Send aggregated player data."""
+        # Get kara status
+        karaoke = models.Karaoke.objects.get_object()
+
         # Get player
-        player = models.Player.get_or_create()
+        player, _ = models.Player.cache.get_or_create(karaoke=karaoke)
 
         # manually update the player timing
         now = datetime.now(tz)
         if player.playlist_entry:
             if not player.paused:
-                player.update(timing=player.timing + (now - player.date), date=now)
+                player.timing += now - player.date
+                player.date = now
 
         # Get player errors
         player_errors_pool = models.PlayerError.objects.all()
-
-        # Get kara status
-        karaoke = models.Karaoke.objects.get_object()
 
         serializer = serializers.DigestSerializer(
             {
@@ -300,8 +302,7 @@ class KaraokeView(drf_generics.RetrieveUpdateAPIView):
             send_to_channel("playlist.device", "send_idle")
 
             # clear player
-            player = models.Player()
-            player.save()
+            player = models.Player.cache.create(karaoke=karaoke)
 
             # empty the playlist
             models.PlaylistEntry.objects.all().delete()
@@ -316,7 +317,7 @@ class KaraokeView(drf_generics.RetrieveUpdateAPIView):
             and "player_play_next_song" in serializer.validated_data
             and karaoke.player_play_next_song
         ):
-            player = models.Player.get_or_create()
+            player, _ = models.Player.cache.get_or_create(karaoke=karaoke)
 
             # request the player to play the next song if idle,
             # and there is a next song to play
@@ -380,10 +381,10 @@ class PlayerStatusView(drf_generics.RetrieveUpdateAPIView):
 
         method = getattr(self, method_name)
 
-        # call the method
-        player.update(**serializer.validated_data)
+        super().perform_update(serializer)
+
+        # call the method and save player
         method(entry, player)
-        player.save()
 
         # broadcast to the front
         # send_to_channel("playlist.front", "send.player.status", {"player": player})
@@ -394,7 +395,7 @@ class PlayerStatusView(drf_generics.RetrieveUpdateAPIView):
         playlist_entry.set_finished()
 
         # reset the player
-        player.reset()
+        player = models.Player.cache.create(pk=player.pk)
 
         # log the info
         logger.debug("The player has finished playing '%s'", playlist_entry)
@@ -408,6 +409,9 @@ class PlayerStatusView(drf_generics.RetrieveUpdateAPIView):
         playlist_entry.set_playing()
         playlist_entry.set_finished()
 
+        # reset the player
+        player = models.Player.cache.create(pk=player.pk)
+
         # log the info
         logger.debug("The player could not play '%s'", playlist_entry)
 
@@ -420,7 +424,9 @@ class PlayerStatusView(drf_generics.RetrieveUpdateAPIView):
         playlist_entry.set_playing()
 
         # update the player
-        player.update(in_transition=True, timing=timedelta(seconds=0))
+        player.in_transition = True
+        player.timing = timedelta(seconds=0)
+        player.save()
 
         # log the info
         logger.debug("Playing transition of entry '%s'", playlist_entry)
@@ -428,7 +434,8 @@ class PlayerStatusView(drf_generics.RetrieveUpdateAPIView):
     def receive_started_song(self, playlist_entry, player):
         """The player started the song of a playlist entry."""
         # update the player
-        player.update(in_transition=False)
+        player.in_transition = False
+        player.save()
 
         # log the info
         logger.debug("Playing song of entry '%s'", playlist_entry)
@@ -436,7 +443,8 @@ class PlayerStatusView(drf_generics.RetrieveUpdateAPIView):
     def receive_paused(self, playlist_entry, player):
         """The player switched to pause."""
         # update the player
-        player.update(paused=True)
+        player.paused = True
+        player.save()
 
         # log the info
         logger.debug("The player switched to pause")
@@ -444,13 +452,16 @@ class PlayerStatusView(drf_generics.RetrieveUpdateAPIView):
     def receive_resumed(self, playlist_entry, player):
         """The player resumed playing."""
         # update the player
-        player.update(paused=False)
+        player.paused = False
+        player.save()
 
         # log the info
         logger.debug("The player resumed playing")
 
     def get_object(self):
-        return models.Player.get_or_create()
+        karaoke = models.Karaoke.objects.get_object()
+        player, _ = models.Player.cache.get_or_create(karaoke=karaoke)
+        return player
 
 
 class PlayerErrorView(drf_generics.ListCreateAPIView):